+++ extralit_ocr/jobs.py
@@ -16,45 +16,9 @@ from extralit_server.models.database import Document
 from rq import get_current_job
 from rq.decorators import job
 
-from extralit_ocr.extract import ExtractionConfig, extract_markdown_with_hierarchy
+from extralit_ocr.extract import ExtractionConfig, _get_document_margins, extract_markdown_with_hierarchy
 
 _LOGGER = logging.getLogger(__name__)
-
-
-def _get_document_margins(metadata: DocumentProcessingMetadata) -> tuple[int, int, int, int] | None:
-    """
-    Fetch margins from document metadata in database.
-
-    Returns:
-        Tuple of (left, top, right, bottom) margins in PDF points, or None if not found
-    """
-    try:
-        if (
-            not metadata.analysis_metadata
-            or not metadata.analysis_metadata.layout_analysis
-            or not metadata.analysis_metadata.layout_analysis.margin_analysis
-        ):
-            print(f"No layout analysis or margin data found in \n{metadata}")
-            return None
-
-        margin_analysis = metadata.analysis_metadata.layout_analysis.margin_analysis
-        print("margin_analysis:", margin_analysis)
-
-        # Convert pixels to PDF points (multiply by 0.75)
-        if all(key in margin_analysis for key in ["left_px", "top_px", "right_px", "bottom_px"]):
-            left = int(margin_analysis["left_px"] * 0.75)
-            top = int(margin_analysis["top_px"] * 0.75)
-            right = int(margin_analysis["right_px"] * 0.75)
-            bottom = int(margin_analysis["bottom_px"] * 0.75)
-
-            margins = (left, top, right, bottom)
-            print(f"Using document-specific margins: {margins}")
-            return margins
-
-    except Exception as e:
-        _LOGGER.warning(f"Error retrieving margins for document: {e}", exc_info=True)
-
-    return None
 
 
 @job(queue=OCR_QUEUE, connection=REDIS_CONNECTION, timeout=900, result_ttl=3600)
